refactor(modules): auth module
