+++ frontend/src/app/app.module.ts
@@ -2,8 +2,6 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
 
-import { JwtModule } from '@auth0/angular-jwt';
-
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 import { HttpClientModule } from "@angular/common/http";
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
@@ -13,9 +11,6 @@ import { AppComponent } from './app.component';
 import { AppConfigModule } from './app-config.module';
 import { AppRoutingModule } from './app-routing.module';
 
-import { tokenGetter } from './services/auth/auth.service';
-import { environment } from '../environments/environment';
-
 @NgModule({
   declarations: [AppComponent],
   entryComponents: [],
@@ -24,13 +19,7 @@ import { environment } from '../environments/environment';
     IonicModule.forRoot(),
     AppConfigModule,
     AppRoutingModule,
-    HttpClientModule,
-    JwtModule.forRoot({
-      config: {
-        tokenGetter: tokenGetter,
-        whitelistedDomains: [ environment.endpoint ]
-      }
-    }),
+    HttpClientModule
   ],
   providers: [
     StatusBar,
