+++ frontend/src/app/auth/auth-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { SigninPage } from './pages/signin/signin.page';
+import { SignupPage } from './pages/signup/signup.page';
+
+const routes: Routes = [
+  { path: '', redirectTo: 'signin', pathMatch: 'full' },
+  { path: 'signin', component: SigninPage },
+  { path: 'signup', component: SignupPage }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class AuthRoutingModule { }
