+++ frontend/src/app/auth/auth.module.spec.ts
@@ -1,0 +1,13 @@
+import { AuthModule } from './auth.module';
+
+describe('AuthModule', () => {
+  let authModule: AuthModule;
+
+  beforeEach(() => {
+    authModule = new AuthModule();
+  });
+
+  it('should create an instance', () => {
+    expect(authModule).toBeTruthy();
+  });
+});
