+++ frontend/src/app/auth/pages/signin/signin.page.html
@@ -13,6 +13,6 @@
     </ion-list>
 
     <ion-button type="submit" [disabled]="!form.valid">Sign in</ion-button>
-    <ion-button color="secondary" [href]="'/signup'">Sign up</ion-button>
+    <ion-button color="secondary" [href]="'/auth/signup'">Sign up</ion-button>
   </form>
 </ion-content>
