+++ frontend/src/app/auth/pages/signin/signin.page.ts
@@ -1,23 +1,19 @@
 import { Component, OnInit } from '@angular/core';
 import { Validators, FormBuilder, FormGroup } from '@angular/forms';
 
-import { NavController } from '@ionic/angular';
-
-import { AuthService } from '../../services';
+import { AuthService } from '../../providers/auth/auth.service';
 
 @Component({
-  selector: 'app-signup',
-  templateUrl: './signup.page.html',
-  styleUrls: ['./signup.page.scss'],
+  selector: 'app-signin',
+  templateUrl: './signin.page.html',
+  styleUrls: ['./signin.page.scss'],
 })
-export class SignupPage implements OnInit {
+export class SigninPage implements OnInit {
   private form: FormGroup;
 
   constructor(private formBuilder: FormBuilder, private authService: AuthService) {
     this.form = this.formBuilder.group(
       {
-        name: ['', Validators.required],
-        lastname: ['', Validators.required],
         email: ['', Validators.required],
         password: ['', Validators.required]
       }
@@ -27,7 +23,7 @@ export class SignupPage implements OnInit {
   ngOnInit() {
   }
 
-  register() {
-    this.authService.register(this.form.value);
+  login() {
+    this.authService.login(this.form.value);
   }
 }
