+++ frontend/src/app/auth/pages/signup/signup.page.html
@@ -24,6 +24,6 @@
 
 
     <ion-button type="submit" [disabled]="!form.valid">Sign up</ion-button>
-    <ion-button color="secondary" [href]="'/signin'">Sign in</ion-button>
+    <ion-button color="secondary" [href]="'/auth/signin'">Sign in</ion-button>
   </form>
 </ion-content>
