--- frontend/src/app/pages/signin/signin.module.ts
@@ -1,27 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { Routes, RouterModule } from '@angular/router';
-
-import { IonicModule } from '@ionic/angular';
-
-import { SigninPage } from './signin.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: SigninPage
-  }
-];
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    IonicModule,
-    RouterModule.forChild(routes)
-  ],
-  declarations: [SigninPage]
-})
-export class SigninPageModule {}
