--- frontend/src/app/pages/signup/signup.module.ts
@@ -1,27 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { Routes, RouterModule } from '@angular/router';
-
-import { IonicModule } from '@ionic/angular';
-
-import { SignupPage } from './signup.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: SignupPage
-  }
-];
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    IonicModule,
-    RouterModule.forChild(routes)
-  ],
-  declarations: [SignupPage]
-})
-export class SignupPageModule {}
